+++ packages/@biorate/command/index.d.ts
@@ -1,0 +1,1 @@
+declare module '@biorate/command' {}