+++ packages/@biorate/mocha-spec/src/spec.ts
@@ -1,15 +1,34 @@
 import * as supertest from 'supertest';
+import * as nock from 'nock';
+import * as sinon from 'sinon';
+import * as chai from 'chai';
+import type { Test, SuperTest } from 'supertest';
+import { jestSnapshotPlugin } from 'mocha-chai-jest-snapshot';
 import { api } from './api';
 import { Unit } from './unit';
 import { Validator } from './validator';
 import { IUnitOptions, IValidatorOptions } from './interfaces';
 
+chai.use(jestSnapshotPlugin());
+
 export abstract class Spec {
+  public static get nock() {
+    return nock;
+  }
+
+  public static get sinon() {
+    return sinon;
+  }
+
+  public static get chai() {
+    return chai;
+  }
+
   protected testDir = 'tests';
 
   protected abstract get httpServer(): any;
 
-  #supertest: supertest.SuperTest<supertest.Test>;
+  #supertest: SuperTest<Test>;
 
   #unit = new Unit(this.testDir);
 
@@ -18,13 +37,25 @@ export abstract class Spec {
     return this.#supertest;
   }
 
+  protected get nock() {
+    return nock;
+  }
+
+  protected get sinon() {
+    return sinon;
+  }
+
+  protected get chai() {
+    return chai;
+  }
+
   protected logReq(method: string, url: string, data: string) {}
 
   protected logRes(status: number, body: string) {}
 
   protected api(url?: string) {
     return api(
-      url ? supertest(url) : this.#supertest,
+      url ? supertest(url) : this.supertest,
       this.logReq.bind(this),
       this.logRes.bind(this),
     );
@@ -36,5 +67,9 @@ export abstract class Spec {
 
   protected validate(options: IValidatorOptions) {
     return Validator.validate(options);
+  }
+
+  protected exactly(result: any, expect: any, message?: string) {
+    return chai.expect(result).is.deep.equal(expect, message);
   }
 }