+++ packages/@biorate/mocha-spec/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.115.0](https://github.com/biorate/core/compare/v1.114.0...v1.115.0) (2025-04-15)
+
+### Features
+
+- **mocha-stub:** exactly options added ([e2357c3](https://github.com/biorate/core/commit/e2357c3410bdb28252acc2a7578f53fb0da5e35f))
+- **mocha-stub:** exactly options added ([d695aed](https://github.com/biorate/core/commit/d695aed0dee56d454876f4638adbd75652ba5622))
+
+## [1.110.2](https://github.com/biorate/core/compare/v1.110.1...v1.110.2) (2025-03-19)
+
+### Bug Fixes
+
+- **mocha-spec:** supertest api method ([63a9a5d](https://github.com/biorate/core/commit/63a9a5d5988e25041d90d24adde3a4cb3e9b5350))
+
+## [1.110.1](https://github.com/biorate/core/compare/v1.110.0...v1.110.1) (2025-03-18)
+
+### Bug Fixes
+
+- **mocha-spec:** deps ([104aea0](https://github.com/biorate/core/commit/104aea0992720546e8a4b9eba95c10b61d4ada7c))
+
+## [1.109.1](https://github.com/biorate/core/compare/v1.109.0...v1.109.1) (2025-03-13)
+
+### Bug Fixes
+
+- **mocha-spec:** types ([6ad4159](https://github.com/biorate/core/commit/6ad415978337dc5e4f0edd149d3b3b2110511512))
+
+# [1.109.0](https://github.com/biorate/core/compare/v1.108.0...v1.109.0) (2025-03-13)
+
+### Features
+
+- **mocha-spec:** deps -> devDeps, chai, nock, sinon, supertest added into devDeps ([6168d1e](https://github.com/biorate/core/commit/6168d1e0b130de25c85bae934ca8b71d464f9d8e))
+
+# [1.108.0](https://github.com/biorate/core/compare/v1.107.0...v1.108.0) (2025-03-13)
+
+### Features
+
+- **mocha-spec:** api test helper added ([29be9e0](https://github.com/biorate/core/commit/29be9e01b66699975c57c58e9cfc7b794895c872))
+
 ## [1.102.1](https://github.com/biorate/core/compare/v1.102.0...v1.102.1) (2025-02-28)
 
 **Note:** Version bump only for package @biorate/mocha-spec